+++ packages/mangler-utils/CHANGELOG.md
@@ -9,11 +9,13 @@ Versioning].
 ## [Unreleased]
 
 - Export character sets. ([#159])
+- Add helper function to handle optional inputs. ([#182])
 
 ## [0.1.23] - 2021-07-23
 
 - Extract _Webmangler Core_ mangler utils into this package.
 
 [#159]: https://github.com/ericcornelissen/webmangler/pull/159
+[#182]: https://github.com/ericcornelissen/webmangler/pull/182
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/ "Keep a CHANGELOG"
 [semantic versioning]: https://semver.org/spec/v2.0.0.html "Semantic versioning"